# docs/swagger.yaml
basePath: /
definitions:
  dto.AppointmentResponse:
    properties:
      created_at:
        type: string
      doctor_id:
        type: string
      duration:
        type: integer
      id:
        type: string
      notes:
        type: string
      patient_id:
        type: string
      reason:
        type: string
      scheduled_at:
        type: string
      service_id:
        type: string
      service_name:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
  dto.CreateAppointmentRequest:
    properties:
      appointment_date:
        example: "2025-11-15"
        type: string
      appointment_time:
        example: "10:00"
        type: string
      doctor_id:
        example: uuid
        type: string
      reason:
        example: Consulta general
        type: string
      service_id:
        example: uuid
        type: string
    type: object
  dto.CreateScheduleRequest:
    properties:
      blocks:
        items:
          $ref: '#/definitions/dto.TimeBlock'
        type: array
      day_of_week:
        example: 1
        type: integer
      doctor_id:
        example: uuid
        type: string
    type: object
  dto.CreateServiceRequest:
    properties:
      description:
        example: Consulta médica general
        type: string
      duration_minutes:
        example: 30
        type: integer
      name:
        example: Consulta General
        type: string
      price:
        example: 80
        type: number
    type: object
  dto.CreateUserRequest:
    properties:
      email:
        example: doctor@clinica.com
        type: string
      first_name:
        example: Juan
        type: string
      last_name:
        example: Pérez
        type: string
      password:
        example: password123
        type: string
      phone:
        example: "+593999999999"
        type: string
      role:
        example: doctor
        type: string
    type: object
  dto.DashboardSummary:
    properties:
      cancellation_rate:
        type: number
      cancelled_appointments:
        type: integer
      completed_appointments:
        type: integer
      confirmed_appointments:
        type: integer
      pending_appointments:
        type: integer
      total_appointments:
        type: integer
      total_doctors:
        type: integer
      total_patients:
        type: integer
      total_revenue:
        type: number
    type: object
  dto.ErrorResponse:
    properties:
      error:
        example: Error message
        type: string
    type: object
  dto.LoginRequest:
    properties:
      email:
        example: admin@clinica.com
        type: string
      password:
        example: admin123
        type: string
    type: object
  dto.LoginResponse:
    properties:
      expires_at:
        type: string
      token:
        type: string
      user:
        $ref: '#/definitions/dto.UserResponse'
    type: object
  dto.RevenueByService:
    properties:
      revenue:
        type: number
      service_id:
        type: string
      service_name:
        type: string
      total_citas:
        type: integer
    type: object
  dto.ScheduleResponse:
    properties:
      blocks:
        items:
          $ref: '#/definitions/dto.TimeBlock'
        type: array
      created_at:
        type: string
      day_of_week:
        type: integer
      doctor_id:
        type: string
      id:
        type: string
    type: object
  dto.ServiceResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      duration_minutes:
        type: integer
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      price:
        type: number
    type: object
  dto.TimeBlock:
    properties:
      end_time:
        example: "13:00"
        type: string
      start_time:
        example: "09:00"
        type: string
    type: object
  dto.TimeSlot:
    properties:
      end_time:
        example: "09:30"
        type: string
      start_time:
        example: "09:00"
        type: string
    type: object
  dto.TopDoctor:
    properties:
      completed_appointments:
        type: integer
      doctor_id:
        type: string
      doctor_name:
        type: string
      total_appointments:
        type: integer
    type: object
  dto.TopService:
    properties:
      service_id:
        type: string
      service_name:
        type: string
      total_citas:
        type: integer
    type: object
  dto.UserResponse:
    properties:
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      role:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: zensoftsas@gmail.com
    name: Zensoft Team
    url: https://github.com/zensoftsas/sistema-reservas
  description: |
    API completa para gestión de clínica médica con sistema de citas, servicios, horarios personalizados y analytics.

    Características principales:
    - Autenticación JWT con roles (admin, doctor, patient)
    - Sistema de citas médicas con validaciones
    - Servicios médicos con precios y duraciones
    - Horarios personalizados por doctor
    - Cálculo dinámico de slots disponibles
    - Analytics y reportes para administradores
    - Notificaciones por email (SendGrid)
    - Recordatorios automáticos
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Sistema de Reservas - Clínica Internacional API
  version: "1.0"
paths:
  /api/analytics/dashboard:
    get:
      consumes:
      - application/json
      description: Retorna resumen con KPIs principales del sistema (solo admin)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DashboardSummary'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Dashboard summary
      tags:
      - Analytics
  /api/analytics/revenue:
    get:
      consumes:
      - application/json
      description: Retorna ingresos agrupados por servicio (solo admin)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.RevenueByService'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Estadísticas de ingresos
      tags:
      - Analytics
  /api/analytics/top-doctors:
    get:
      consumes:
      - application/json
      description: Retorna ranking de doctores por cantidad de citas (solo admin)
      parameters:
      - default: 10
        description: Número de doctores a retornar
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.TopDoctor'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Top doctores más solicitados
      tags:
      - Analytics
  /api/analytics/top-services:
    get:
      consumes:
      - application/json
      description: Retorna ranking de servicios por demanda (solo admin)
      parameters:
      - default: 10
        description: Número de servicios a retornar
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.TopService'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Top servicios más populares
      tags:
      - Analytics
  /api/appointments:
    post:
      consumes:
      - application/json
      description: Crear una nueva cita médica con validaciones de disponibilidad
      parameters:
      - description: Datos de la cita
        in: body
        name: appointment
        required: true
        schema:
          $ref: '#/definitions/dto.CreateAppointmentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.AppointmentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Crear cita médica
      tags:
      - Appointments
  /api/appointments/my:
    get:
      consumes:
      - application/json
      description: Retorna todas las citas del usuario autenticado
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.AppointmentResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Obtener mis citas
      tags:
      - Appointments
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Autenticar usuario y obtener token JWT
      parameters:
      - description: Email y password
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Login de usuario
      tags:
      - Authentication
  /api/schedules:
    post:
      consumes:
      - application/json
      description: Crear horario personalizado para un doctor (solo admin)
      parameters:
      - description: Datos del horario
        in: body
        name: schedule
        required: true
        schema:
          $ref: '#/definitions/dto.CreateScheduleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ScheduleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Crear horario de doctor
      tags:
      - Schedules
  /api/schedules/doctor/{id}:
    get:
      consumes:
      - application/json
      description: Retorna todos los horarios configurados para un doctor
      parameters:
      - description: ID del doctor
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.ScheduleResponse'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Obtener horarios de un doctor
      tags:
      - Schedules
...
  /api/services/available-slots:
    get:
      consumes:
      - application/json
      description: Calcula y retorna los slots de tiempo disponibles para un doctor,
        servicio y fecha específicos
      parameters:
      - description: ID del usuario doctor
        in: query
        name: doctor_id
        required: true
        type: string
      - description: ID del servicio
        in: query
        name: service_id
        required: true
        type: string
      - description: Fecha en formato YYYY-MM-DD
        in: query
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.TimeSlot'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Obtener horarios disponibles
      tags:
      - Services
  /api/services/create:
    post:
      consumes:
      - application/json
      description: Crear un nuevo servicio médico (solo admin)
      parameters:
      - description: Datos del servicio
        in: body
        name: service
        required: true
        schema:
          $ref: '#/definitions/dto.CreateServiceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ServiceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Crear servicio
      tags:
      - Services
  /api/users:
    post:
      consumes:
      - application/json
      description: Crear un nuevo usuario (admin, doctor o patient)
      parameters:
      - description: Datos del usuario
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Crear usuario
      tags:
      - Users
  /api/users/me:
    get:
      consumes:
      - application/json
      description: Retorna información del usuario actual basado en el token JWT
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Obtener perfil del usuario autenticado
      tags:
      - Users
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...